+++ crates/macros/cgp-macro-core/src/types/cgp_data/mod.rs
@@ -1,5 +1,7 @@
+mod derive_builder;
 mod derive_has_fields;
 mod record;
 
+pub use derive_builder::*;
 pub use derive_has_fields::*;
 pub use record::*;
