+++ crates/macros/cgp-macro-lib/src/derive_extractor/extract_field_impls.rs
@@ -1,8 +1,8 @@
+use cgp_macro_core::types::cgp_data::{index_to_generic_ident, to_generic_args};
 use cgp_macro_core::types::field::Symbol;
 use quote::quote;
 use syn::{Arm, GenericArgument, Ident, ItemEnum, ItemImpl, Type, parse2};
 
-use crate::derive_builder::{index_to_generic_ident, to_generic_args};
 use crate::derive_extractor::get_variant_type;
 
 pub fn derive_extract_field_impls(
