+++ crates/macros/cgp-macro-lib/src/derive_extractor/extractor_enum.rs
@@ -1,8 +1,8 @@
+use cgp_macro_core::types::cgp_data::index_to_generic_ident;
 use proc_macro2::Span;
 use quote::quote;
 use syn::{GenericParam, Ident, ItemEnum, Lifetime, LifetimeParam, Type, TypeParam, parse2};
 
-use crate::derive_builder::index_to_generic_ident;
 use crate::derive_extractor::{get_variant_type, type_to_variant_fields};
 
 pub fn derive_extractor_enum(
