+++ crates/macros/cgp-macro-lib/src/derive_extractor/partial_data.rs
@@ -1,7 +1,6 @@
+use cgp_macro_core::types::cgp_data::index_to_generic_ident;
 use quote::quote;
 use syn::{Ident, ItemEnum, ItemImpl, parse2};
-
-use crate::derive_builder::index_to_generic_ident;
 
 pub fn derive_partial_data_impl(
     context_struct: &ItemEnum,
