+++ crates/macros/cgp-macro-lib/src/entrypoints/cgp_record.rs
@@ -1,4 +1,4 @@
-use cgp_macro_core::types::cgp_data::{ItemCgpRecord, derive_has_fields_impls_from_struct};
+use cgp_macro_core::types::cgp_data::ItemCgpRecord;
 use proc_macro2::TokenStream;
 use quote::quote;
 use syn::{ItemStruct, parse2};
@@ -14,7 +14,7 @@ pub fn derive_cgp_record_from_struct(item_struct: ItemStruct) -> syn::Result<Tok
     let record = ItemCgpRecord { item_struct };
 
     let has_field_impls = record.to_has_field_impls()?;
-    let has_fields_impls = derive_has_fields_impls_from_struct(&record.item_struct)?;
+    let has_fields_impls = record.to_has_fields_impls()?;
     let build_field_impls = derive_build_field_from_struct(&record.item_struct)?;
 
     Ok(quote! {
